[lexical-react][lexical-table] Chore: Optimize table observers
